requirements.txt
asgiref==3.4.1
boto3==1.20.5
botocore==1.23.5
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.1.13
django-phonenumber-field==5.2.0
django-storages==1.12.3
django-twilio==0.13.1.post0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
phonenumbers==8.12.35
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
South==1.0.2
sqlparse==0.4.1
twilio==6.63.2
urllib3==1.26.7
whitenoise==5.3.0
